# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the main API server for the Calendar Reg application.",
        "title": "Calendar Reg Main API",
        "contact": {
            "name": "API Support",
            "email": "[EMAIL_ADDRESS]"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/courses": {
            "get": {
                "description": "Retrieve a paginated list of courses. Use limit=0 to fetch all.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Get courses (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 10, 0=all)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create a new course with sections and schedules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Create a new course",
                "parameters": [
                    {
                        "description": "Course Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CourseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/courses/{code}": {
            "get": {
                "description": "Retrieve a specific course details by its code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Get course by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CourseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Soft delete a course (set deleted_at timestamp)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Soft delete course by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CourseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "credits": {
                    "type": "string"
                },
                "faculty": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name_en": {
                    "type": "string"
                },
                "name_th": {
                    "type": "string"
                },
                "prerequisite": {
                    "type": "string"
                },
                "program": {
                    "type": "string"
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SectionResponse"
                    }
                },
                "semester": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateCourseRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "credits": {
                    "type": "string"
                },
                "faculty": {
                    "type": "string"
                },
                "name_en": {
                    "type": "string"
                },
                "name_th": {
                    "type": "string"
                },
                "prerequisite": {
                    "type": "string"
                },
                "program": {
                    "type": "string"
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SectionRequest"
                    }
                },
                "semester": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ScheduleRequest": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.ScheduleResponse": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.SectionRequest": {
            "type": "object",
            "properties": {
                "exam_date": {
                    "type": "string"
                },
                "instructor": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ScheduleRequest"
                    }
                },
                "seats": {
                    "type": "integer"
                }
            }
        },
        "dto.SectionResponse": {
            "type": "object",
            "properties": {
                "exam_end": {
                    "type": "string"
                },
                "exam_start": {
                    "type": "string"
                },
                "instructor": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ScheduleResponse"
                    }
                },
                "seats": {
                    "type": "integer"
                }
            }
        },
...
    }
}